requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-Login==0.6.3
Werkzeug==2.3.7
google-api-python-client==2.100.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.0.0
psycopg[binary]
bleach==6.1.0
Flask-WTF==1.2.1
email-validator==2.1.1
openpyxl==3.1.2
pandas==2.0.3
xlsxwriter==3.1.9
weasyprint==60.1
reportlab==4.0.7
